refactor(editor): move history from doc to store (#12131)
