+++ blocksuite/affine/blocks/embed-doc/src/embed-synced-doc-block/configs/toolbar.ts
@@ -310,7 +310,7 @@ const builtinSurfaceToolbarConfig = {
               note.props.displayMode !== NoteDisplayMode.EdgelessOnly
           );
 
-        ctx.doc.captureSync();
+        ctx.store.captureSync();
         ctx.chain
           .pipe(duplicateSelectedModelsCommand, {
             selectedModels: [model],
@@ -350,7 +350,7 @@ const builtinSurfaceToolbarConfig = {
         }
         if (contentModels.length === 0) return;
 
-        ctx.doc.captureSync();
+        ctx.store.captureSync();
         ctx.chain
           .pipe(draftSelectedModelsCommand, {
             selectedModels: contentModels,
