+++ packages/frontend/core/src/modules/workspace/impls/doc.ts
@@ -9,7 +9,6 @@ import {
   type Workspace,
   type YBlock,
 } from '@blocksuite/affine/store';
-import { signal } from '@preact/signals-core';
 import { Subject } from 'rxjs';
 import { Awareness } from 'y-protocols/awareness.js';
 import * as Y from 'yjs';
@@ -21,10 +20,6 @@ type DocOptions = {
 };
 
 export class DocImpl implements Doc {
-  private readonly _canRedo = signal(false);
-
-  private readonly _canUndo = signal(false);
-
   private readonly _collection: Workspace;
 
   private readonly _storeMap = new Map<string, Store>();
@@ -32,13 +27,6 @@ export class DocImpl implements Doc {
   // doc/space container.
   private readonly _handleYEvents = (events: Y.YEvent<YBlock | Y.Text>[]) => {
     events.forEach(event => this._handleYEvent(event));
-  };
-
-  private _history!: Y.UndoManager;
-
-  private readonly _historyObserver = () => {
-    this._updateCanUndoRedoSignals();
-    this.slots.historyUpdated.next();
   };
 
   private readonly _initSubDoc = () => {
@@ -55,9 +43,7 @@ export class DocImpl implements Doc {
       }
     }
 
-    const spaceDoc = new Y.Doc({
-      guid: this.id,
-    });
+    const spaceDoc = new Y.Doc({ guid: this.id });
     spaceDoc.clientID = this.rootDoc.clientID;
     this._loaded = false;
 
@@ -71,19 +57,6 @@ export class DocImpl implements Doc {
 
   /** Indicate whether the block tree is ready */
   private _ready = false;
-
-  private _shouldTransact = true;
-
-  private readonly _updateCanUndoRedoSignals = () => {
-    const canRedo = this._history.canRedo();
-    const canUndo = this._history.canUndo();
-    if (this._canRedo.peek() !== canRedo) {
-      this._canRedo.value = canRedo;
-    }
-    if (this._canUndo.peek() !== canUndo) {
-      this._canUndo.value = canUndo;
-    }
-  };
 
   protected readonly _yBlocks: Y.Map<YBlock>;
 
@@ -103,8 +76,6 @@ export class DocImpl implements Doc {
 
   readonly slots = {
     // eslint-disable-next-line rxjs/finnish
-    historyUpdated: new Subject<void>(),
-    // eslint-disable-next-line rxjs/finnish
     yBlockUpdated: new Subject<
       | {
           type: 'add';
@@ -123,20 +94,8 @@ export class DocImpl implements Doc {
     return this.workspace.blobSync;
   }
 
-  get canRedo() {
-    return this._canRedo.peek();
-  }
-
-  get canUndo() {
-    return this._canUndo.peek();
-  }
-
   get workspace() {
     return this._collection;
-  }
-
-  get history() {
-    return this._history;
   }
 
   get isEmpty() {
@@ -217,19 +176,6 @@ export class DocImpl implements Doc {
   private _initYBlocks() {
     const { _yBlocks } = this;
     _yBlocks.observeDeep(this._handleYEvents);
-    this._history = new Y.UndoManager([_yBlocks], {
-      trackedOrigins: new Set([this._ySpaceDoc.clientID]),
-    });
-
-    this._history.on('stack-cleared', this._historyObserver);
-    this._history.on('stack-item-added', this._historyObserver);
-    this._history.on('stack-item-popped', this._historyObserver);
-    this._history.on('stack-item-updated', this._historyObserver);
-  }
-
-  /** Capture current operations to undo stack synchronously. */
-  captureSync() {
-    this._history.stopCapturing();
   }
 
   clear() {
@@ -250,7 +196,6 @@ export class DocImpl implements Doc {
 
   dispose() {
     this._destroy();
-    this.slots.historyUpdated.unsubscribe();
 
     if (this.ready) {
       this._yBlocks.unobserveDeep(this._handleYEvents);
@@ -335,45 +280,8 @@ export class DocImpl implements Doc {
     return this;
   }
 
-  redo() {
-    this._history.redo();
-  }
-
-  undo() {
-    this._history.undo();
-  }
-
   remove() {
     this._destroy();
     this.rootDoc.getMap('spaces').delete(this.id);
-  }
-
-  resetHistory() {
-    this._history.clear();
-  }
-
-  /**
-   * If `shouldTransact` is `false`, the transaction will not be push to the history stack.
-   */
-  transact(fn: () => void, shouldTransact: boolean = this._shouldTransact) {
-    this._ySpaceDoc.transact(
-      () => {
-        try {
-          fn();
-        } catch (e) {
-          console.error(
-            `An error occurred while Y.doc ${this._ySpaceDoc.guid} transacting:`
-          );
-          console.error(e);
-        }
-      },
-      shouldTransact ? this.rootDoc.clientID : null
-    );
-  }
-
-  withoutTransact(callback: () => void) {
-    this._shouldTransact = false;
-    callback();
-    this._shouldTransact = true;
   }
 }
